adventure_guide/requirements.txt
boto3
fastapi
httpx
python-dotenv
python-multipart
uvicorn[standard]
grpcio
grpclib 
grpcio-tools
protobuf
